initial commit for collection view feature
